refactor: extract highlightPreserve, usePasteAlignFix, useEditorFocusRecovery from useEditorSetup
useEditorSetup.js shrinks from ~655 lines to ~175 lines. Each extracted
module owns one concern: highlight mark preservation across edits,
paste alignment/depth/highlight repair, and three focus-recovery
behaviours (touch guard, desktop deep-link click, selectionchange).
